pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-model-testbed"
version = "0.1.0"
description = "A minimal, clean structure to test different AI models behind one interface"
authors = [{name = "AI Testbed Team"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "colorama>=0.4.0",
    "requests>=2.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
]

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["src"]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "*Tests"